Tidy up

+++ src/hbin_ele.cpp
@@ -9,8 +9,11 @@
 #include "window.h"
 
 
-hbin_element::hbin_element(int n_x1, int n_y1, int n_x2, int n_y2, int n_z, window & n_owner, hbin & n_hb, int n_scroll_x, int n_scroll_y, int n_selection):
-  element(n_x1, n_y1, n_x2, n_y2, n_z, n_owner), m_hb(n_hb), m_scroll_x(n_scroll_x), m_scroll_y(n_scroll_y), m_selection(n_selection)
+hbin_element::hbin_element(int16_t n_x1, int16_t n_y1, int16_t n_x2,
+    int16_t n_y2, int16_t n_z, window & n_owner, hbin & n_hb,
+    int16_t n_scroll_x, int16_t n_scroll_y, int16_t n_selection)
+  : element(n_x1, n_y1, n_x2, n_y2, n_z, n_owner), m_hb(n_hb),
+    m_scroll_x(n_scroll_x), m_scroll_y(n_scroll_y), m_selection(n_selection)
 {
 }
 
@@ -18,11 +21,12 @@ hbin_element::hbin_element(int n_x1, int n_y1, int n_x2, int n_y2, int n_z, wind
 void hbin_element::paint(graphics & g) const
 {
   g.draw_bitmap(-m_scroll_x, -m_scroll_y, m_hb.background());
-  g.draw_bitmap(-m_scroll_x + m_hb.link_x(m_selection), -m_scroll_y + m_hb.link_y(m_selection), m_hb.link_picture(m_selection));
+  g.draw_bitmap(-m_scroll_x + m_hb.link_x(m_selection),
+    -m_scroll_y + m_hb.link_y(m_selection), m_hb.link_picture(m_selection));
 }
 
 
-void hbin_element::set_scroll_pos(int x, int y)
+void hbin_element::set_scroll_pos(int16_t x, int16_t y)
 {
   m_scroll_x = x;
   m_scroll_y = y;
@@ -33,12 +37,12 @@ void hbin_element::set_scroll_pos(int x, int y)
 }
 
 
-void hbin_element::set_selection(int n_selection)
+void hbin_element::set_selection(int16_t n_selection)
 {
-  int x1;
-  int y1;
-  int x2;
-  int y2;
+  int16_t x1;
+  int16_t y1;
+  int16_t x2;
+  int16_t y2;
 
   assert(n_selection < m_hb.num_links());
   if(n_selection == m_selection) return;
